Don't link to a one-element listing page from recent updates
If a package is built as a split package where pkgname != pkgbase, but
only one actual split package is produced, the link on the recent update
screen requires an extra click to get to the single package. Fix this by
linking directly to the package itself.

(Examples in current repos: ntfs-3g, python2-south)

Signed-off-by: Dan McGee <dan@archlinux.org>