Move stats portion of developer dashboard to separate view
This stuff is all below the fold when the page first loads, and adds a
good amount of loading time to the developer dashboard. Split it out,
where it will be wired back and hooked up via an AJAX insertion in a
future commit. Both parts work standalone as is in this commit.

Signed-off-by: Dan McGee <dan@archlinux.org>