Add mirror error logs to mirror details page
Give a window of 7 days for logs here rather than the default 24 hours
we do on the main status page since we are only retrieving details for a
single mirror with a handful of URLs. This should make it easier to have
all information regarding one mirror in a single location.

Signed-off-by: Dan McGee <dan@archlinux.org>