reporead: bring back batched_bulk_create()
For packages with filelists with > 80,000 items, we were starting to see
some serious memory issues in reporead. This was both on the statement
generation side in Python as well as on the database side. Break the
updates into chunks of 10,000 when we encounter packages with tons of
files to control things in a bit.

Signed-off-by: Dan McGee <dan@archlinux.org>