Set all attributes to default values on status URL fetch
We were missing two duration-related attributes here, causing some 500
errors to happen if we had cached status_data around that didn't agree
with our current list of checked mirrors. Don't blow up on the JSON data
fetch by ensuring we provide a value, even if it is out of date.

Signed-off-by: Dan McGee <dan@archlinux.org>