Break out developer reports into a separate module
This code was getting quite unwieldy, and wasn't very modular. Introduce
a DeveloperReport class that contains the content for a single report,
and utilize it to create our various report metadata and package
filtering operations. Utilize these report objects in the reports view,
vastly simplifying it.

We don't yet dynamically generate the list of reports on the developer
index page; that will be coming soon.

Signed-off-by: Dan McGee <dan@archlinux.org>