Add true parameter to Jinja2 default filter
We need this to make it act more like the Django filter, which prints
the default value not only when the variable is undefined, but also when
it evaluates to being false (such as a None value).

Noticed by felixonmars on packages that provided other packages; None
was always showing as the comparison operator. More embarrassing is that
our file lists were always showing 'None' concatenated onto all
directory entries.

Signed-off-by: Dan McGee <dan@archlinux.org>