package details: remove category from bugtracker links
This completely screws up searching for bugs, as you can easily miss any
bugs filed to the testing package, and will *always* miss the ones that
have been moved to "Upstream Bugs".

There is no purpose in restricting the view of bugs filed against a
specific package, to some arbitrary subcategory of "why" the bug was
filed; all it accomplishes is causing unknowing users to file duplicate
bug reports.