Misc fixes (#84)
* Update coveragerc to exclude newly located tests

There are now tests under main/tests/* etc, which should be omitted for
coverage.

* omit settings from coverage

settings and local_settings aren't interesting for coverage, omit them.

* remove leftover secure context processor

The secure context processor was used to determine if the cdn served
assets should be loaded over https or http. Since assets are no longer
served with a cdn and the whole site is loaded over https these days,
this code is dead and can be removed.

* packages: Add test for flagging packages out of date

Add simple test cases for flagging packages out of date, should later be
improved to query the Package model for packages and use that data
instead of hardcoding.