Fix Python 2 to 3 Unicode issues for string join
Revisited earlier commit where email subject lines with potentially
mixed encoding are joined into a single string. This fix brings in
the 'codecs' import to decode bytes to string using either the passed
encoding (if provided) or otherwise utf-8.

Changed the test_donor_import to no longer convert the Header to a
string and instead leave it as a byte array.