main: Handle multiple testing repositories
Since there are multiple [testing] repositories when a package is in
kde-unstable, testing and flagged in extra the in_testing function will
generate an error since there are multiple results. Rewrite the query so
it handles multiple results and order it by id since [testing] should
have a lower id then kde-unstable.