Refactor CSS colours
The colours have been refactored into CSS variables,
which also sport a dark mode variant.

Colours are largely the same as before in light mode,
with the exception of similar-ish colours used for the
same purpose being coalesced into a single variable and thus colour,
e.g. all the various colours in the #bbb-#ccc range
being used for borders becoming part of the --borders
variable.

Signed-off-by: Janet Blackquill <uhhadd@gmail.com>