Use Release models torrent data for feeds
When a new Release is done, the iso and torrent file might not be
rsync'd yet to the archweb server, which makes torrent clients which
watch the feed unable to retrieve the torrent file and give up. As this
behaviour is unwanted use the Release model torrent data which is always
availabe.