Fix PXE netboot image names
templates/releng/archlinux.ipxe:
Fix the naming of the initrd images.
In a commit leading to archiso 46 [1] (image 2020-08-01) the way files
are being copied in the build script has been made more generic. Due to
this change the names of the provided PXE netboot images on the download
servers have changed as well and stopped matching the files as provided
in archweb's .ipxe template.
Add amd-ucode.img as additional initird to be downloaded, verified and
used.

[1] https://gitlab.archlinux.org/archlinux/archiso/-/commit/47e11125e4672fc3448caefa6728439160daedd3