Add Tier 0 Mirror Access support
Support generating an access token for access to our tier0 mirror and
other services which are behind a nginx using
ngx_http_auth_request_module for authentication. This works as
following, nginx requires authentication for a route which is protected
and redirects the request to archweb which handles the authentication.

Closes: #347