Allow users to set their own names
Users should be able to set their own names as they see fit, since names
change for any number of reasons, and there's no sense in allowing them
to be registered with whatever name they give but then require DevOps to
update it if it for whatever reason needs to be updated.

This will hopefully also lead to a better visibility of the latin name
field, allowing staff members whose names are properly written in
non-latin scripts to have their names properly represented.

Signed-off-by: Johannes Löthberg <johannes@kyriasis.com>