templates/public/download.html: update checksum and signature list and verification
Link to *sum.txt files on archlinux.org, just like with the GPG
signatures, as the mirrors cannot be trusted.

Add the GnuPG verification command from https://wiki.archlinux.org/title/Installation_guide#Verify_signature.
It would be better if we do not duplicate it in both archweb and the wiki.