Simplify instructions when using GnuPG for signature verification
Since Pierre's signatures contain [Signer's UID packet][UID] GnuPG
will [automatically fetch the signing key][T3324] using WKD when using
`--auto-key-retrieve`. The filename is assumed to be the same as the
signature name minus the `.sig` suffix. Ask the user to confirm that
the primary key's fingerprint is correct which makes sure that a
correct signing key has been used and not any other that may be in
their keyring.

[UID]: https://www.rfc-editor.org/rfc/rfc4880#section-5.2.3.22
[T3324]: https://dev.gnupg.org/T3324